Refactor search and filtering logic into model extension functions
This change moves the search and filtering logic from a centralized utility class to specific extension functions within the API model files, improving encapsulation and readability.

Included changes:
- Deleted `FilterUtils.kt` and migrated its logic to `search` extension functions for `List<Event>`, `List<Movie>`, and `List<Timer>`.
- Updated `TvViewModel`, `RadioViewModel`, `MoviesViewModel`, `TimersViewModel`, and various EPG ViewModels to use the new extension functions.
- Centralized `ContentFlag.Channel` filtering for events within the `Event.search` extension to ensure consistent search results across different screens.
- Standardized search behavior to return `null` when filter terms are blank or the source list is empty, simplifying null-safety checks in `combine` blocks.
- Maintained the use of `Dispatchers.Default` within the extension functions to ensure heavy filtering and scoring operations remain off the main thread.
